requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.51.0
supabase==2.7.4
python-dotenv==1.0.1
anthropic==0.34.2
rich==13.8.1
